requirements.txt
# Core dependencies
requests>=2.31.0
aiohttp>=3.9.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# AI Integration
google-generativeai>=0.3.0
anthropic>=0.18.0

# Web scraping and automation
playwright>=1.40.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Caching
redis>=5.0.0
hiredis>=2.2.0

# AWS Storage
boto3>=1.34.0
botocore>=1.34.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0

# Async support
asyncio-mqtt>=0.16.0
aiofiles>=23.2.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3
urllib3>=2.1.0
certifi>=2023.7.0

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
pytest-cov>=4.1.0
black>=23.12.0
flake8>=7.0.0
mypy>=1.8.0
pre-commit>=3.6.0

# Testing utilities
responses>=0.24.0
httpx>=0.26.0
fakeredis>=2.20.0
moto>=4.2.0